backend/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 데이터베이스 관련
sqlalchemy>=2.0.23
asyncmy>=0.2.9
aiomysql>=0.2.0
alembic>=1.12.1

# Pydantic 및 설정 관리
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 벡터 데이터베이스
qdrant-client>=1.7.1

# AI API Services  
# openai>=1.3.7  # OpenAI API (마이그레이션 후 제거 예정)
google-generativeai>=0.3.0  # Google Gemini API

# 비동기 HTTP 클라이언트
httpx>=0.25.2
aiohttp>=3.9.1

# 인증 및 보안
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 환경 변수 관리
python-dotenv>=1.0.0

# 로깅 및 모니터링
structlog>=23.2.0
rich>=13.7.0

# 데이터 처리
pandas>=2.1.4
numpy>=1.25.2

# 날짜/시간 처리
python-dateutil>=2.8.2

# JSON 처리
orjson>=3.9.10

# 이미지 처리 (프로필 이미지용)
Pillow>=10.1.0

# 텍스트 처리
nltk>=3.8.1
# konlpy>=0.6.0  # 한국어 처리 (필요시 주석 해제)

# 캐싱 (선택사항)
# redis>=5.0.1
# aioredis>=2.0.1

# 테스트
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# 개발 도구
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
# mypy>=1.7.1  # 타입 체크 (필요시 주석 해제)

# 배포 및 컨테이너
gunicorn>=21.2.0

# 기타 유틸리티
click>=8.1.7
jinja2>=3.1.2
markupsafe>=2.1.3 